+++ apple-nvram/Cargo.toml
@@ -13,5 +13,6 @@ repository = "https://github.com/WhatAmISupposedToPutHere/asahi-nvram"
 [dependencies]
 adler32 = "1"
 crc32fast = "1.3.2"
+either = "1.9.0"
 indexmap = "2.0.0"
 nix = "0.25"
